PYTHON-6014 Move internal pool classes into pool_shared.py
