+++ test/asynchronous/utils.py
@@ -30,11 +30,12 @@ from inspect import iscoroutinefunction
 
 from bson.son import SON
 from pymongo import AsyncMongoClient
-from pymongo.asynchronous.pool import Pool, _CancellationContext, _PoolGeneration
+from pymongo.asynchronous.pool import Pool
 from pymongo.errors import ConfigurationError
 from pymongo.hello import HelloCompat
 from pymongo.lock import _async_create_lock
 from pymongo.operations import _Op
+from pymongo.pool_shared import _CancellationContext, _PoolGeneration
 from pymongo.read_preferences import ReadPreference
 from pymongo.server_selectors import any_server_selector, writable_server_selector
